+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blocksummary/FinalizationReward.java
@@ -1,7 +1,7 @@
 package com.concordium.sdk.responses.blocksummary;
 
 import com.concordium.sdk.transactions.AccountAddress;
-import com.concordium.sdk.transactions.GTUAmount;
+import com.concordium.sdk.transactions.CCDAmount;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -12,14 +12,14 @@ import java.util.Objects;
 @Getter
 @ToString
 public final class FinalizationReward {
-    private GTUAmount amount;
+    private CCDAmount amount;
     private AccountAddress address;
 
     @JsonCreator
     FinalizationReward(@JsonProperty("amount") String amount,
                        @JsonProperty("address") String address) {
         if (!Objects.isNull(amount)) {
-            this.amount = GTUAmount.fromMicro(amount);
+            this.amount = CCDAmount.fromMicro(amount);
         }
         if (!Objects.isNull(address)) {
             this.address = AccountAddress.from(address);